Added some missing dialog code.  Note that this is a framework only and not a fully working set of dialogs.  It allows someone to start filling in the bits of code needed to properly run the dialogs.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@2020 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/beos/ap_BeOSFrame.cpp
@@ -274,7 +274,7 @@ void AP_BeOSFrame::setXScrollRange(void)
 	int width = ((AP_FrameData*)m_pData)->m_pDocLayout->getWidth();
 	be_Window *pBWin = (be_Window*)getTopLevelWindow();
         pBWin->Lock();
-        int windowWidth = pBWin->m_pbe_DocView->Bounds().Width();
+        int windowWidth = (int)pBWin->m_pbe_DocView->Bounds().Width();
         pBWin->Unlock();
         BScrollBar *hscroll = pBWin->m_hScroll;
 
@@ -315,7 +315,7 @@ void AP_BeOSFrame::setYScrollRange(void)
 	int height = ((AP_FrameData*)m_pData)->m_pDocLayout->getHeight();
 	be_Window *pBWin = (be_Window*)getTopLevelWindow();
         pBWin->Lock();
-        int windowHeight = pBWin->m_pbe_DocView->Bounds().Height();
+        int windowHeight = (int)pBWin->m_pbe_DocView->Bounds().Height();
         pBWin->Unlock();
         BScrollBar *vscroll = pBWin->m_vScroll;
